Add updated copies of project files
The project files included in the `googletest` submodule that pertain to
projects that will be needed for this repository need to have
modifications made to them in order to be usable. This commit creates
copies of those project files that were created by Visual Studio 2017 by
opening the gtest-md solution and the gmock solution and allowing VS to
convert the project files to the appropriate version. The newly create
project files were then moved out of the submodule, so they can be
managed separately from it.

A future commit will modify these project files to update the files and
directories they reference to be accurate (since the project files were
moved, breaking the references), and to modify other features of them.